% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{adv_sales}
\alias{adv_sales}
\title{Monthly advertising and sales data}
\format{
A data frame with 36 rows and 3 variables
}
\usage{
data(adv_sales)
}
\description{
Monthly advertising and sales data
}
\details{
Description provided in attr(adv_sales, "description")
}
\keyword{datasets}
