% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{Arima}
\alias{Arima}
\title{ARIMA model}
\usage{
Arima(
  dataset,
  rvar,
  order = c(1, 0, 0),
  seasonal = NULL,
  evar = NULL,
  mean = TRUE,
  drift = FALSE,
  model = NULL,
  data_filter = "",
  envir = parent.frame(),
  ...
)
}
\arguments{
\item{dataset}{Dataset}

\item{rvar}{The response variable in the model}

\item{order}{Order ...}

\item{seasonal}{Seasonal order ...}

\item{evar}{Additional explanatory variables in the model}

\item{mean}{Mean ...}

\item{drift}{Drift ...}

\item{model}{Model ...}

\item{data_filter}{Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")}

\item{envir}{Environment to extract data from}

\item{...}{other parameters to pass to forecast::Arima}
}
\value{
A list of all inputs and results as an object of class Arima
}
\description{
ARIMA model
}
\details{
See \url{https://radiant-rstats.github.io/docs/ets/arima.html} for an example in Radiant
}
\examples{
Arima(diamonds, "price", order = c(1, 0, 1))
}
\seealso{
\code{\link{summary.Arima}} to summarize results

\code{\link{plot.Arima}} to plot results

\code{\link{forecast.Arima}} to generate predictions
}
