% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{plot.arima.predict}
\alias{plot.arima.predict}
\title{Plot method for model.predict functions}
\usage{
\method{plot}{arima.predict}(
  x,
  xvar = "",
  facet_row = ".",
  facet_col = ".",
  color = "none",
  conf_lev = 0.95,
  ...
)
}
\arguments{
\item{x}{Return value from predict functions (e.g., predict.Arima)}

\item{xvar}{Variable to display along the X-axis of the plot}

\item{facet_row}{Create vertically arranged subplots for each level of the selected factor variable}

\item{facet_col}{Create horizontally arranged subplots for each level of the selected factor variable}

\item{color}{Adds color to a scatter plot to generate a heat map. For a line plot one line is created for each group and each is assigned a different color}

\item{conf_lev}{Confidence level to use for prediction intervals (.95 is the default)}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for model.predict functions
}
\seealso{
\code{\link{predict.Arima}} to generate predictions
}
