% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant.ets_window}
\alias{radiant.ets_window}
\title{Launch radiant.ets in an Rstudio window}
\usage{
radiant.ets_window(state, ...)
}
\arguments{
\item{state}{Path to state file to load}

\item{...}{additional arguments to pass to shiny::runApp (e.g, port = 8080)}
}
\description{
Launch radiant.ets in an Rstudio window
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant.ets_window()
}
}
