% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{store.arima.predict}
\alias{store.arima.predict}
\title{Store predicted values from an Arima model}
\usage{
\method{store}{arima.predict}(dataset, object, name = "prediction", ...)
}
\arguments{
\item{dataset}{Dataset to add predictions to}

\item{object}{Return value from model predict function}

\item{name}{Variable name(s) assigned to predicted values}

\item{...}{Additional arguments}
}
\description{
Store predicted values from an Arima model
}
\details{
See \url{https://radiant-rstats.github.io/docs/ets/arima.html} for an example in Radiant
}
\examples{
Arima(diamonds, rvar = "price") \%>\%
  Arima(diamonds, model = .) \%>\%
  store(diamonds, ., name = "pred_pref")

}
